Introduce `Nightmare_service.Parser`
Currently, only a really naive parser has been implemented and it allows
to extract the anchor, the query-string and the fragments of a path. It
replaces the `sanitize_path` of `Muhokama`. Currently, extracting the
query-string and anchor will probably be of little use, but it is to
ensure that a path is built on the relevant elements.