Introduce `Nightmare_service.Path`
A `Path` is a URL fragment. It is essentially a component of an endpoint
allowing bi-directional use (link generation and URL interpretation).
I have taken the liberty of compartmentalising each element of a service
into different, testable modules. The implementation is quite similar to
the Muhokama paths, the main difference is in:
- The documentation
- The terminology used (`sprintf` and `sscanf`) instead of `make` and
  `handle` links because because I find that the notion of links comes
  "too early" in the hierarchy of levels that constitute a service.
- Typed variables can be defined by the user using the `variable inj
  proj name` function.

Co-authored-by: xhtmlboi <xhtmlboi@gmail.com>
Co-authored-by: xvw <xaviervdw@gmail.com>