Introduce Nightmare_service toplevel and `~:`
Dealing with local openings to describe paths can be quickly boring
hence the introduction of a "toplevel" module that re-exports the
definitions usual infix operators.
Later, it is possible that this module will include the contents of the
Service module (so that `Nightmare.Service = Nightmare_service`).

The commit reintroduces `~:`, as in Muhokama except that its type is set
to act only on functions taking `unit` and returning an endpoint.