Introduce `Nightmare_common` and `Error`
The commit introduces `Nightmare_common` to list utilities common to
different packages in the "Nightmare" umbrella and error registration
support. The implementation, suggested by @xvw, is very strongly
inspired by that of Tezos, which provides imperative combiners to record
errors. This allows handlers to be handled locally and not force errors
relating to, for example, form handling, into the common library. Each
module can "instantiate its set of errors".

As, at the time of this contribution, Preface still does not support the
functor/applicative/monad triad with arity 2 types, using polymorphic
variants would have made interaction with Try complicated. Furthermore,
the big advantage, from my point of view, is that it allows not to have
a complete knowledge of the errors to be reported and each error is
interested, individually, in its pretty-printing strategy.

Co-authored-by: xvw <xaviervdw@gmail.com>