Some renaming and use also modules for var
As the type parameters are relatively long, this commit reduces
them (from `handler_continuation` and `handler_return` to `continuation`
and `witness`.

In addition, we add the possibility of defining a variable in a [Path]
using a module.

And it fix the date of the license since we take a lot of time...