Prohibition of parameters captured by an endpoint
Before this commit, it was possible to write this kind of
code (supposing that `tyxml.html` and `nightmare_tyxml` are available in the
current scope):

```ocaml
let my_link =
  a_of ~a:[ a_href "https://google.be"]
     my_endpoint my_param [txt "Link"]
```

Which is quite annoying because the `a_href` attribute overwrites the
one derived from the provisioned endpoint!

Fortunately, @Octachron, through @xvw, has presented an approach to
prohibit certain polymorphic variant constructors. For example, in the
case of `a` attributes (`Html_types.a_attrib`):

```ocaml
type a_attrib =
  [ | common | `Href | `Hreflang | `Media | `Rel | `Target | `Mime_type
    | `Download
  ]
```

We redefine a type that excludes the `Href`:

```ocaml
type a_attrib_without_href =
  [ | common | Hreflang | `Media | `Rel | `Target | `Mime_type
    | `Download
  ]
```

And now, we can define a list **prohibiting** receiving `href`, but
allowing to be _cons_ with an `a_href` using **these coercion logic**:

```ocaml
let my_a_attrib ?(a = []) my_href =
  a_href my_href
  :: ( a
        : [< a_attrib_without_href ] attrib list
        :> [> Html_types.a_attrib] attrib list)
```

This makes the use of these functions more robust!

Co-authored-by: xvw <xaviervdw@gmail.com>