Move Interfaces and Bindings outside of the API
Although it's likely that we'll discover a new form of convention later,
at the moment we seem to be used to having a `Bindings` module that
describes the low-level plumbing of a JS library.

Re-exporting it from the root of the module, especially in Nightmare_js,
creates conflicts with potential `Bindings` modules present in a library
dependent on `Nightmare_js`.